requirements.txt
google-adk==1.2.1
pandas==2.2.3
requests==2.32.3
flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
python-dotenv==1.1.0
streamlit==1.44.1
pytest==8.3.4
pytest-cov==6.0.0
